Operational Alpha
The core pipeline should be set up and possible to run on a new system.
I will test it on Genesis with Qsub commands and xhost machines without qsub
to see how it works. Also will write steps neccesary to set up on a new
system and add it to readme 'installation'

Good luck!